:truck: Move some components around to try to come up with a standard naming pattern.
